refactor: migrate MediaRemoteDataSource to ProcessGateway
- Rename MediaRemoteBridge → MediaRemoteDataSourceImpl (consistent naming)
- Replace internal Process/pipe management with gateway.runStreaming
- Script deployment (ensureScript) remains in DataSource; only subprocess
  execution is delegated to gateway
- DI registration updated

Part of #196
