+++ Package.swift
@@ -310,6 +310,7 @@ let package = Package(
             name: "MediaRemoteDataSource",
             dependencies: [
                 "Domain",
+                .product(name: "Dependencies", package: "swift-dependencies"),
                 .product(name: "Files", package: "Files"),
             ],
             resources: [.copy("Resources/media-remote-helper.swift")]
